Feature/nc 1212 (Fix zabbix host creation/deletion )

 - move backend from folder to file.
 - allow to define hosts templates in NC templates.
 - Delete host only after proper resource deletion.

See merge request !27

+++ src/nodeconductor_zabbix/handlers.py
@@ -1,5 +1,7 @@
 import logging
 
+from nodeconductor.structure.models import Resource
+
 from .models import Host
 
 
@@ -13,15 +15,11 @@ def update_hosts_visible_name_on_scope_name_change(sender, instance, **kwargs):
         backend.update_visible_name(host)
 
 
-def delete_hosts_on_scope_deletion(sender, instance, **kwargs):
-    for host in Host.objects.filter(scope=instance):
-        if host.backend_id:
-            try:
+def delete_hosts_on_scope_deletion(sender, instance, name, source, target, **kwargs):
+    if target == Resource.States.DELETING:
+        for host in Host.objects.filter(scope=instance):
+            if host.backend_id:
                 backend = host.get_backend()
                 backend.destroy(host)
-            except Exception as e:
-                logger.error(
-                    'Zabbix host "%s" (UUID: %s) deletion fails with exception %s', host.name, host.uuid.hex, e)
+            else:
                 host.delete()
-        else:
-            host.delete()
